Add Java code for formatting date and time using Calendar
This commit adds a Java class named Cal that demonstrates how to format date and time using the Calendar class. It includes various examples of formatting date and time strings using different conversion characters. This code can be used as a reference for working with date and time in Java.

#java #datetime #calendar